✨ feat: add event correlation for validator change detection and pool deltas
Implements event-based change detection for validators using lifecycle events
(AutoPause, AutoBeginUnstaking, Slash, Reward) to catch state transitions even
when RPC fields haven't updated yet. This solves the race condition where events
fire before the RPC state reflects the change.

Changes:
- Add event correlation to ConvertValidatorsWithChangeDetection with 3-phase detection:
  1. Check for lifecycle events (state transitions)
  2. Check for new validators
  3. Compare RPC fields (only if no event occurred)
- Parse validator lifecycle events and map by address for O(1) lookup
- Add H-1 delta calculations to convertPools (amount, total points, LP count)
- Calculate derived pool IDs (committee-based addressing)
- Fix database schema: change SMALLINT to INTEGER for pool columns (chain_id, lp_count, lp_count_delta)
- Update add-canopy-node.sh to support both node-1 and node-2 registration
- Update docker-compose.blockchain.yml port mappings for node-2 (avoid conflicts)

+++ pkg/db/postgres/chain/pool.go
@@ -11,10 +11,10 @@ func (db *DB) initPools(ctx context.Context) error {
 		CREATE TABLE IF NOT EXISTS pools (
 			pool_id INTEGER NOT NULL,                      -- UInt32 -> INTEGER
 			height BIGINT NOT NULL,
-			chain_id SMALLINT NOT NULL,                    -- UInt16 -> SMALLINT (renamed from pool_chain_id)
+			chain_id INTEGER NOT NULL,                    -- UInt16 -> INTEGER (renamed from pool_chain_id)
 			amount BIGINT NOT NULL DEFAULT 0,
 			total_points BIGINT NOT NULL DEFAULT 0,
-			lp_count SMALLINT NOT NULL DEFAULT 0,          -- UInt16 -> SMALLINT
+			lp_count INTEGER NOT NULL DEFAULT 0,          -- UInt16 -> INTEGER
 			height_time TIMESTAMP WITH TIME ZONE NOT NULL,
 			liquidity_pool_id INTEGER NOT NULL DEFAULT 0,  -- UInt32 -> INTEGER
 			holding_pool_id INTEGER NOT NULL DEFAULT 0,    -- UInt32 -> INTEGER
@@ -22,7 +22,7 @@ func (db *DB) initPools(ctx context.Context) error {
 			reward_pool_id INTEGER NOT NULL DEFAULT 0,     -- UInt32 -> INTEGER
 			amount_delta BIGINT NOT NULL DEFAULT 0,        -- Int64 -> BIGINT
 			total_points_delta BIGINT NOT NULL DEFAULT 0,  -- Int64 -> BIGINT
-			lp_count_delta SMALLINT NOT NULL DEFAULT 0,    -- Int16 -> SMALLINT
+			lp_count_delta INTEGER NOT NULL DEFAULT 0,    -- Int16 -> INTEGER
 			PRIMARY KEY (pool_id, height)
 		);
 
@@ -42,7 +42,7 @@ func (db *DB) initPoolPointsByHolder(ctx context.Context) error {
 			pool_id INTEGER NOT NULL,                      -- UInt32 -> INTEGER
 			height BIGINT NOT NULL,
 			height_time TIMESTAMP WITH TIME ZONE NOT NULL,
-			committee SMALLINT NOT NULL,                   -- UInt16 -> SMALLINT
+			committee INTEGER NOT NULL,                   -- UInt16 -> INTEGER
 			points BIGINT NOT NULL DEFAULT 0,
 			liquidity_pool_points BIGINT NOT NULL DEFAULT 0,
 			liquidity_pool_id INTEGER NOT NULL DEFAULT 0,  -- UInt32 -> INTEGER
